[ci][microcheck/11] move get_changed_tests to Test class
