+++ ci/ray_ci/test_tester.py
@@ -17,8 +17,6 @@ from ci.ray_ci.tester import (
     _get_flaky_test_targets,
     _get_tag_matcher,
     _get_new_tests,
-    _get_changed_files,
-    _get_changed_tests,
     _get_human_specified_tests,
 )
 from ray_release.test import Test, TestState
@@ -128,7 +126,7 @@ def test_get_test_targets() -> None:
             "ray_release.test.Test.gen_high_impact_tests",
             return_value={"step": test_objects},
         ), mock.patch(
-            "ci.ray_ci.tester._get_changed_tests",
+            "ray_release.test.Test.get_changed_tests",
             return_value=set(),
         ), mock.patch(
             "ci.ray_ci.tester._get_new_tests",
@@ -256,7 +254,7 @@ def test_get_high_impact_test_targets() -> None:
             "ci.ray_ci.tester._get_new_tests",
             return_value=test["new_tests"],
         ), mock.patch(
-            "ci.ray_ci.tester._get_changed_tests",
+            "ray_release.test.Test.get_changed_tests",
             return_value=test["changed_tests"],
         ), mock.patch(
             "ci.ray_ci.tester._get_human_specified_tests",
@@ -283,30 +281,6 @@ def test_get_new_tests(mock_gen_from_s3, mock_run_script_with_output) -> None:
     assert _get_new_tests(
         "linux", LinuxTesterContainer("test", skip_ray_installation=True)
     ) == {"//new_test"}
-
-
-@mock.patch.dict(
-    os.environ,
-    {"BUILDKITE_PULL_REQUEST_BASE_BRANCH": "base", "BUILDKITE_COMMIT": "commit"},
-)
-@mock.patch("subprocess.check_call")
-@mock.patch("subprocess.check_output")
-def test_get_changed_files(mock_check_output, mock_check_call) -> None:
-    mock_check_output.return_value = b"file1\nfile2\n"
-    assert _get_changed_files() == {"file1", "file2"}
-
-
-@mock.patch("ci.ray_ci.tester._get_test_targets_per_file")
-@mock.patch("ci.ray_ci.tester._get_changed_files")
-def test_get_changed_tests(
-    mock_get_changed_files, mock_get_test_targets_per_file
-) -> None:
-    mock_get_changed_files.return_value = {"test_src", "build_src"}
-    mock_get_test_targets_per_file.side_effect = (
-        lambda x: {"//t1", "//t2"} if x == "test_src" else {}
-    )
-
-    assert _get_changed_tests() == {"//t1", "//t2"}
 
 
 @mock.patch.dict(
